Minor fix in show_end_game_dialog function in core.c
Fixed call to the `move` function during the placement of the dialog
text where the vertical position was incremented prior to the text
placement such that it was 1 row below its intended position.

The issue has now been fixed and the vertical position is now incemented
after the movement has been performed.